backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
gunicorn==23.0.0
aiofiles==24.1.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
python-dotenv==1.0.1
openai==1.58.1
anthropic==0.40.0
fpdf2==2.8.1
cryptography==43.0.3
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
